+++ .env
@@ -1,8 +1,9 @@
 #########################
 # Job Server
 #########################
-API_HOST="http://localhost:5000"
-API_SERVER_PORT=5000
+# App config
+JOB_APP_PORT=5000
+JOB_APP_ENVIROMENT=development
 
 
 #########################
@@ -32,6 +33,7 @@ DB_NAME=internado
 DB_DIALECT=postgres
 DB_HOST=postgres
 DB_PORT=5432
+DB_RESTART=always
 
 # TypeORM-specific
 TYPEORM_SYNCHRONIZE=true
@@ -44,3 +46,10 @@ TYPEORM_ENTITIES=dist/models/**/*.js
 ########################
 WAIT_SLEEP_INTERVAL=60
 WAIT_HOSTS_TIMEOUT=600
+
+
+########################
+# Adminer
+########################
+ADMINER_PORT=8080
+ADMINER_RESTART=always
