+++ auth_server/package.json
@@ -1,0 +1,59 @@
+{
+    "name": "internado-auth-server",
+    "version": "0.0.1",
+    "description": "Find personalized internship opportunities based on your skills and interests",
+    "main": "server.js",
+    "scripts": {
+        "build": "npm run clean && tsc -p .",
+        "lint": "tslint -p tsconfig.json -c tslint.json",
+        "dev": "./node_modules/.bin/tsc-watch -p . --onSuccess \"node dist/server.js\"",
+        "start": "npm run build && npm run lint && node dist/server.js",
+        "prepare": "npm run build && npm run lint",
+        "clean": "rimraf dist",
+        "test": "echo \"Error: no test specified\" && exit 1"
+    },
+    "cpu": [
+        "x64"
+    ],
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/ubclaunchpad/Internado.git"
+    },
+    "author": "UBC - Launchpad",
+    "license": "MIT",
+    "bugs": {
+        "url": "https://github.com/ubclaunchpad/Internado/issues"
+    },
+    "homepage": "https://github.com/ubclaunchpad/Internado#readme",
+    "dependencies": {
+        "@types/bcrypt": "^3.0.0",
+        "@types/body-parser": "^1.17.0",
+        "@types/express": "^4.16.0",
+        "@types/jsonwebtoken": "^8.3.0",
+        "@types/morgan": "^1.7.35",
+        "@types/node": "^8.10.39",
+        "@types/passport": "^1.0.0",
+        "@types/passport-jwt": "^3.0.1",
+        "@types/pg": "^7.4.11",
+        "@types/validator": "^10.9.0",
+        "bcrypt": "^3.0.4",
+        "body-parser": "^1.18.3",
+        "concurrently": "^4.1.0",
+        "debug": "^4.1.0",
+        "dotenv": "^6.1.0",
+        "express": "^4.16.4",
+        "jsonwebtoken": "^8.4.0",
+        "morgan": "^1.9.1",
+        "passport": "^0.4.0",
+        "passport-jwt": "^4.0.0",
+        "pg": "^7.3.0",
+        "reflect-metadata": "^0.1.10",
+        "ts-node": "^3.3.0",
+        "tsc-watch": "^2.1.2",
+        "tslint": "^5.11.0",
+        "typeorm": "0.2.9",
+        "typescript": "^3.3.3",
+        "validator": "^10.11.0",
+        "rimraf": "^2.6.2"
+    }
+}
