+++ auth_server/src/configurations/app.ts
