+++ auth_server/src/configurations/jwt.ts
