+++ auth_server/src/routes/v1.ts
