+++ auth_server/tsconfig.json
