+++ job_server/.dockerignore
@@ -1,3 +1,3 @@
-**/node_modules
+node_modules
 npm-debug.log
 ormconfig.json
