+++ job_server/tsconfig.json
@@ -4,6 +4,8 @@
             "es5",
             "es6"
         ],
+        "outDir": "./dist/",
+        "declaration": false,
         "target": "es5",
         "module": "commonjs",
         "moduleResolution": "node",
@@ -12,8 +14,8 @@
         "experimentalDecorators": true,
         "sourceMap": true
     },
-    "filesGlob": [
-        "/**/*.ts"
+    "include": [
+      "**/*.ts"
     ],
     "typeRoots": [
         "./node_modules/@types"
@@ -21,4 +23,4 @@
     "exclude": [
         "node_modules"
     ]
-}
+  }
