--- server/package.json
@@ -1,49 +1,0 @@
-{
-    "name": "server",
-    "version": "0.0.1",
-    "description": "Find personalized internship opportunities based on your skills and interests",
-    "main": "server.js",
-    "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1",
-        "build": "tsc -p . && tslint -p tsconfig.json -c tslint.json",
-        "start": "npm run build && ts-node server.ts",
-        "clean": "rimraf **/*.js **/*.js.map node_modules"
-    },
-    "cpu": [
-        "x64"
-    ],
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/ubclaunchpad/Internado.git"
-    },
-    "author": "UBC - Launchpad",
-    "license": "MIT",
-    "bugs": {
-        "url": "https://github.com/ubclaunchpad/Internado/issues"
-    },
-    "homepage": "https://github.com/ubclaunchpad/Internado#readme",
-    "dependencies": {
-        "@types/swagger-ui-express": "^3.0.0",
-        "@types/node": "^8.0.29",
-        "@types/body-parser": "^1.17.0",
-        "@types/express": "^4.16.0",
-        "@types/morgan": "^1.7.35",
-        "@types/pg": "^7.4.11",
-        "tslint": "^5.11.0",
-        "ts-node": "3.3.0",
-        "body-parser": "^1.18.3",
-        "debug": "^4.1.0",
-        "dotenv": "^6.1.0",
-        "express": "^4.16.4",
-        "morgan": "^1.9.1",
-        "parse-error": "^0.2.0",
-        "pg": "^7.3.0",
-        "reflect-metadata": "^0.1.10",
-        "swagger-ui-express": "^4.0.2",
-        "typeorm": "0.2.9",
-        "typescript": "^3.1.3"
-    },
-    "devDependencies": {
-        "rimraf": "^2.6.2"
-    }
-}
